requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
Django==5.0.3
django-crontab==0.7.1
djangorestframework==3.15.0
idna==3.6
pycparser==2.21
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
validators==0.23.2
